lambda/custom/requirements.txt
ask-sdk-core==1.13.0
ask-sdk-dynamodb-persistence-adapter==1.13.0
ask-sdk-model==1.22.0
ask-sdk-runtime==1.13.0
ask-sdk==1.13.0
requests==2.23.0
pytz==2019.3
dataclasses-json